% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/jsonedit.R
\name{jsonedit}
\alias{jsonedit}
\title{listviewer htmlwidget using \href{https://github.com/josdejong/jsoneditor}{jsoneditor}}
\usage{
jsonedit(message, width = NULL, height = NULL)
}
\description{
\code{jsoneditor} provides a flexible and helpful interactive tree-like view of \code{lists}
  or really any R dataset that can be represented as \code{JSON}.
  Eventually, this could become a very nice way to not only view but also modify R data using
  Shiny.
}

