% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit_gadget.R
\name{jsonedit_gadget}
\alias{jsonedit_gadget}
\title{Shiny Gadget for 'jsonedit'}
\usage{
jsonedit_gadget(..., height = NULL, width = NULL)
}
\arguments{
\item{...}{arguments for \code{\link{jsonedit}}}

\item{height, width}{any valid \code{CSS} size unit for the
height and width of the gadget}
}
\description{
Provides a \href{http://shiny.rstudio.com/articles/gadgets.html}{Shiny gadget}
  interface for \code{jsonedit} to interactively edit and return the
  changes for use in R.
}
\examples{
\dontrun{
library(listviewer)

jsonedit_gadget(
  structure(
    as.list(1:4),
    names=letters[1:4]
  )
)
}
}
