% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit_gadget.R
\name{jsoneditAddin}
\alias{jsoneditAddin}
\title{jsoneditAddin}
\usage{
jsoneditAddin()
}
\description{
This uses the gadget as an addin, so you can select objects then
  activate the addin in RStudio either via menu or a hotkey.
}
\keyword{internal}
