% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonedit.R
\name{jsoneditOutput}
\alias{jsoneditOutput}
\title{Widget output function for use in Shiny}
\usage{
jsoneditOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

