% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listviewer-package.r
\docType{package}
\name{listviewer}
\alias{listviewer}
\title{listviewer.}
\description{
htmlwidget for interactive views of R lists
}
\details{
R lists, especially nested lists, can be very difficult to visualize or
represent. \code{str} just isn't enough, so this suite of htmlwidgets is
designed to help see, understand, and maybe even modify your R lists.
}
