% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{number_unnamed}
\alias{number_unnamed}
\title{Number Starting at \code{1}}
\usage{
number_unnamed(l)
}
\arguments{
\item{l}{\code{list}}
}
\description{
JavaScript starts at \code{0}, but R starts at \code{1}.
This means unnamed lists and vectors are indexed starting at
\code{0} in listviewer widgets.  This little helper function
tries to resolve the disconnect by assigning sequential numbers
starting at \code{1} to names for unnamed \code{lists} and \code{vectors}.
Please note though that using \code{number_unnamed} will potentially
cause difficulties serializing back and forth between JavaScript and R.
}
\examples{
library(listviewer)
jsonedit(
  number_unnamed(list(x=list(letters[1:3])))
)

}
