% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recurse}
\alias{recurse}
\title{Recurse Over Nested Lists}
\usage{
recurse(l, func, ...)
}
\arguments{
\item{l}{\code{list} over which we would like to recurse}

\item{func}{\code{function} to apply to the list}

\item{...}{arguments for the \code{func}}
}
\description{
Like \code{rapply} but recurses over nested lists and
retain \code{attributes} and \code{names}.
}
\keyword{internal}
